pyproject.toml
[project]
name = "fte"
version = "1.0.0"
description = "Digital FTE - AI-powered email automation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0",
    "cryptography>=41.0.0",
    "requests>=2.31.0",
    "playwright>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
fte = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
